zsh: Change the out of the glo alias
Color is now only on informational fields. Why? Since I often pipe the
glo alias with the command yank, the color codes get in the clipboard.
They must then be removed before using whatever was put in the clipboard
as color codes are useless when I simply want to have the string itself
to use it in another command